block: Use 64-bit integers explicitly.
In most builds, uint is a 64-bit value, but I prefer to be explicit about the
size since the semantics may otherwise change, and it's important for hashes to
be interoperable across builds.

+++ block/splitter.go
@@ -155,7 +155,7 @@ func (s *Splitter) Next() ([]byte, error) {
 		isCut := false
 		for ; i < s.next; i++ {
 			u := s.hash.Update(s.buf[i])
-			isCut = u%uint(s.exp) == 1 && i-s.end >= s.min
+			isCut = u%uint64(s.exp) == 1 && i-s.end >= s.min
 			if isCut {
 				break
 			}
